visualization of isolation window in identification view
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@7638 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/VISUAL/AxisTickCalculator.C
@@ -64,7 +64,7 @@ namespace OpenMS
 		DoubleReal sDecPow = floor(log10(dx));
 		DoubleReal sDec = pow(10.0,sDecPow);
 
-    UInt n_max_big_gridlines = floor(dx / sDec);
+    UInt n_max_big_gridlines = (UInt)floor(dx / sDec);
 
 		std::vector<DoubleReal> big;
 		DoubleReal currGL = ceilDecimal(x1, (UInt)sDecPow);
